+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2021-09-26
+### Fixed
+- Fixed Context Blocks
+
 ## [1.0.0] - 2021-09-25
-- Added Function for Context
-- Added Function for It
-- Added Function for Handler
+### Added
+- Function for Context
+- Function for It
+- Function for Handler